profile by design guides

+++ widgets/src/commonMain/kotlin/dev/icerock/moko/widgets/style/background/Background.kt
@@ -5,16 +5,15 @@
 package dev.icerock.moko.widgets.style.background
 
 data class Background(
-    val colors: List<Int> = listOf(0xFFFFFFFF.toInt(), 0xFFFFFFFF.toInt()),
-    val colorsDisabled: List<Int> = colors,
+    val colors: List<Long> = listOf(0xFFFFFFFF, 0xFFFFFFFF),
+    val colorsDisabled: List<Long> = colors,
     val direction: Direction = Direction.LEFT_RIGHT,
     val shape: Shape = Shape()
 ) {
     constructor(
-        color: Int,
-        colorDisabled: Int = color,
+        color: Long,
+        colorDisabled: Long = color,
         direction: Direction = Direction.LEFT_RIGHT,
         shape: Shape = Shape()
     ) : this(listOf(color, color), listOf(colorDisabled, colorDisabled), direction, shape)
 }
-
